Revert "Move event and eventbus to shardingsphere-common module (#36274)"
This reverts commit 48574c725b687bffa3336ea5b7d89f6163f8f777.
