+++ infra/util/src/main/java/org/apache/shardingsphere/infra/util/eventbus/EventBusContext.java
@@ -15,23 +15,32 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.event;
+package org.apache.shardingsphere.infra.util.eventbus;
 
-import org.apache.shardingsphere.infra.event.fixture.EventSubscriberFixture;
-import org.junit.jupiter.api.Test;
+import com.google.common.eventbus.EventBus;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-class EventBusContextTest {
+/**
+ * Event bus context.
+ */
+public final class EventBusContext {
     
-    @Test
-    void assertEventBusContextTest() {
-        EventBusContext eventBusContext = new EventBusContext();
-        EventSubscriberFixture listener = new EventSubscriberFixture();
-        eventBusContext.register(listener);
-        eventBusContext.post("foo_event");
-        assertThat(listener.getEvents().size(), is(1));
-        assertThat(listener.getEvents().get(0), is("foo_event"));
+    private final EventBus eventBus = new EventBus();
+    
+    /**
+     * Register event subscriber.
+     *
+     * @param subscriber event subscriber
+     */
+    public void register(final EventSubscriber subscriber) {
+        eventBus.register(subscriber);
+    }
+    
+    /**
+     * Post event.
+     *
+     * @param event event
+     */
+    public void post(final Object event) {
+        eventBus.post(event);
     }
 }
