+++ infra/util/src/test/java/org/apache/shardingsphere/infra/util/eventbus/EventBusContextTest.java
@@ -15,10 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.event.deliver;
+package org.apache.shardingsphere.infra.util.eventbus;
 
-/**
- * Deliver event.
- */
-public interface DeliverEvent {
+import org.apache.shardingsphere.infra.util.eventbus.fixture.EventSubscriberFixture;
+import org.junit.jupiter.api.Test;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+class EventBusContextTest {
+    
+    @Test
+    void assertEventBusContextTest() {
+        EventBusContext eventBusContext = new EventBusContext();
+        EventSubscriberFixture listener = new EventSubscriberFixture();
+        eventBusContext.register(listener);
+        eventBusContext.post("foo_event");
+        assertThat(listener.getEvents().size(), is(1));
+        assertThat(listener.getEvents().get(0), is("foo_event"));
+    }
 }
