+++ mode/core/src/main/java/org/apache/shardingsphere/mode/deliver/DeliverEventSubscriberRegistry.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.mode.deliver;
 
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.infra.event.EventBusContext;
+import org.apache.shardingsphere.infra.util.eventbus.EventBusContext;
 
 import java.util.Collection;
